requirements.txt
# Core
langchain>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0

# Image and GIF handling
Pillow>=10.0.0
imageio>=2.30.0
imageio-ffmpeg>=0.5.0

# Optional utilities
regex>=2023.6.3

# CLI / env
click>=8.1.0
